pyproject.toml
[tool.poetry]
name = "cognizantmlops"
version = "0.1.0"
description = ""
authors = ["Rubén Chuliá Mena <rubchume@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
jupyterlab = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
